debugging elastic client
