+++ src/components/search/SearchBar.jsx
@@ -3,16 +3,13 @@ import useSearchRoute from '@/hooks/useSearchRoute'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'next-i18next'
 import LanguageFilters from '@/components/search/LanguageFilters'
-import { IconGlobe } from '../Icons'
-import { IconCircleArrowRight } from '../Icons'
+
 
 const SearchBar = ({ search }) => {
   // Sync search field with URL
   const { t } = useTranslation('common')
   const [qw, setQuery] = useState(search)
-  const [show, setShow] = useState(false)
   const goToSearch = useSearchRoute({ search: qw })
-  const toggle = () => setShow(!show)
   useEffect(() => {
     setQuery(search)
   }, [search, setQuery])
@@ -34,26 +31,7 @@ const SearchBar = ({ search }) => {
         <button className="inline-flex flex-none items-center w-12 h-12">
           <IconLookingGlass className="mx-2" />
         </button>
-        <div className="flex absolute -bottom-24 lg:-bottom-10 left-0 flex-wrap items-center">
-          <span className="inline-block">
-            <IconGlobe
-              className="w-6 h-6 me-2"
-              title={t('languageMenu.label')}
-              disabled
-            />
-          </span>
-          {!show && (
-            <button type="button" onClick={toggle}>
-              <IconCircleArrowRight className=" w-5 h-5" />
-            </button>
-          )}
-          {show && <LanguageFilters />}
-          {show && (
-            <button type="button" onClick={toggle}>
-              <IconCircleArrowRight className="w-5 h-5 rotate-180" />
-            </button>
-          )}
-        </div>
+      <LanguageFilters />
       </form>
     </div>
   )
